Admin part fully functional and permissions added

+++ src/controllers/ActivitiesController.java
@@ -7,10 +7,8 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
-import model.Activity;
+import model.*;
 import model.Process;
-import model.Task;
-import model.User;
 
 import java.io.IOException;
 
@@ -95,9 +93,17 @@ public class ActivitiesController {
      */
     @FXML
     void createActivityBegining(ActionEvent event) throws IOException {
-        if (selectedProcess != null) {
-            main.openCreateActivity(selectedProcess, signedUser, 1);
+        if (signedUser.getPermission() == Permissions.EDIT) {
+            if (selectedProcess != null) {
+                main.openCreateActivity(selectedProcess, signedUser, 1);
+            }
+        }else {
+            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+            alert.setTitle("Atención");
+            alert.setContentText("No cuenta con los permisos para hacer esto");
+            alert.showAndWait();
         }
+
 
     }
 
@@ -107,9 +113,17 @@ public class ActivitiesController {
      */
     @FXML
     void createActivityEnd(ActionEvent event) throws IOException {
-        if (selectedProcess != null) {
-            main.openCreateActivity(selectedProcess, signedUser, 0);
+        if (signedUser.getPermission() == Permissions.EDIT){
+            if (selectedProcess != null) {
+                main.openCreateActivity(selectedProcess, signedUser, 0);
+            }
+        }else {
+            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+            alert.setTitle("Atención");
+            alert.setContentText("No cuenta con los permisos para hacer esto");
+            alert.showAndWait();
         }
+
 
     }
 
@@ -121,7 +135,15 @@ public class ActivitiesController {
      */
     @FXML
     void createTaks(ActionEvent event) throws IOException {
-        main.openCreateTask(signedUser, selectedProcess, selectedActivity);
+        if (signedUser.getPermission() == Permissions.EDIT){
+            main.openCreateTask(signedUser, selectedProcess, selectedActivity);
+
+        }else {
+            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+            alert.setTitle("Atención");
+            alert.setContentText("No cuenta con los permisos para hacer esto");
+            alert.showAndWait();
+        }
 
     }
 
@@ -131,16 +153,24 @@ public class ActivitiesController {
      */
     @FXML
     void deleteActivity(ActionEvent event) {
-        if(selectedActivity.getTasksList().size() == 0){
-            boolean deleteActivity = singleton.deleteActivity(selectedProcess, selectedActivity);
-            if (deleteActivity) listaActividades.remove(selectedActivity);
-            activityTable.refresh();
+        if (signedUser.getPermission() == Permissions.EDIT){
+            if(selectedActivity.getTasksList().size() == 0){
+                boolean deleteActivity = singleton.deleteActivity(selectedProcess, selectedActivity);
+                if (deleteActivity) listaActividades.remove(selectedActivity);
+                activityTable.refresh();
+            }else {
+                Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                alert.setTitle("Atención");
+                alert.setContentText("Para marcar como completada la actividad no debe tener ninguna tarea");
+                alert.showAndWait();
+            }
         }else {
             Alert alert = new Alert(Alert.AlertType.INFORMATION);
             alert.setTitle("Atención");
-            alert.setContentText("Para marcar como completada la actividad no debe tener ninguna tarea");
+            alert.setContentText("No cuenta con los permisos para hacer esto");
             alert.showAndWait();
         }
+
 
     }
 
@@ -151,34 +181,58 @@ public class ActivitiesController {
      */
     @FXML
     void editActivity(ActionEvent event) throws IOException {
-        if (selectedActivity!= null){
-            main.openEditActivity(selectedActivity, selectedProcess, signedUser);
+        if (signedUser.getPermission() == Permissions.EDIT){
+            if (selectedActivity!= null){
+                main.openEditActivity(selectedActivity, selectedProcess, signedUser);
+            }
+        }else {
+            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+            alert.setTitle("Atención");
+            alert.setContentText("No cuenta con los permisos para hacer esto");
+            alert.showAndWait();
         }
+
 
     }
 
     @FXML
     void editTask(ActionEvent event) throws IOException {
-        if (selectedTask !=null){
-            main.openEditTask(selectedTask, selectedProcess, signedUser);
+        if (signedUser.getPermission() == Permissions.EDIT){
+            if (selectedTask !=null){
+                main.openEditTask(selectedTask, selectedProcess, signedUser);
 
+            }
+        }else {
+            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+            alert.setTitle("Atención");
+            alert.setContentText("No cuenta con los permisos para hacer esto");
+            alert.showAndWait();
         }
+
 
 
     }
 
     @FXML
     void markAsDone(ActionEvent event) {
-        if (singleton.isNextTask(selectedActivity, selectedTask)){
-            boolean task = singleton.markTaskAsDone(selectedActivity, selectedTask);
-            if (task) listaTareas.remove(selectedTask);
-            taskTable.refresh();
-        }else {
+        if (signedUser.getPermission() == Permissions.EDIT){
+            if (singleton.isNextTask(selectedActivity, selectedTask)){
+                boolean task = singleton.markTaskAsDone(selectedActivity, selectedTask);
+                if (task) listaTareas.remove(selectedTask);
+                taskTable.refresh();
+            }else {
+                Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                alert.setTitle("Atención");
+                alert.setContentText("Para marcar como completada una tarea debe estar completada la que la precede");
+                alert.showAndWait();
+            }
+        }else{
             Alert alert = new Alert(Alert.AlertType.INFORMATION);
             alert.setTitle("Atención");
-            alert.setContentText("Para marcar como completada una tarea debe estar completada la que la precede");
+            alert.setContentText("No cuenta con los permisos para hacer esto");
             alert.showAndWait();
         }
+
 
     }
